+++ src/Symfony/Component/Security/Http/EventListener/IsGrantedAttributeListener.php
@@ -12,6 +12,8 @@
 namespace Symfony\Component\Security\Http\EventListener;
 
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+use Symfony\Component\ExpressionLanguage\Expression;
+use Symfony\Component\ExpressionLanguage\ExpressionLanguage;
 use Symfony\Component\HttpKernel\Event\ControllerArgumentsEvent;
 use Symfony\Component\HttpKernel\Exception\HttpException;
 use Symfony\Component\HttpKernel\KernelEvents;
@@ -28,7 +30,8 @@ use Symfony\Component\Security\Http\Attribute\IsGranted;
 class IsGrantedAttributeListener implements EventSubscriberInterface
 {
     public function __construct(
-        private AuthorizationCheckerInterface $authChecker,
+        private readonly AuthorizationCheckerInterface $authChecker,
+        private ?ExpressionLanguage $expressionLanguage = null,
     ) {
     }
 
@@ -42,21 +45,15 @@ class IsGrantedAttributeListener implements EventSubscriberInterface
         $arguments = $event->getNamedArguments();
 
         foreach ($attributes as $attribute) {
-            $subjectRef = $attribute->subject;
             $subject = null;
 
-            if ($subjectRef) {
+            if ($subjectRef = $attribute->subject) {
                 if (\is_array($subjectRef)) {
-                    foreach ($subjectRef as $ref) {
-                        if (!\array_key_exists($ref, $arguments)) {
-                            throw new RuntimeException(sprintf('Could not find the subject "%s" for the #[IsGranted] attribute. Try adding a "$%s" argument to your controller method.', $ref, $ref));
-                        }
-                        $subject[$ref] = $arguments[$ref];
+                    foreach ($subjectRef as $refKey => $ref) {
+                        $subject[\is_string($refKey) ? $refKey : (string) $ref] = $this->getIsGrantedSubject($ref, $arguments);
                     }
-                } elseif (!\array_key_exists($subjectRef, $arguments)) {
-                    throw new RuntimeException(sprintf('Could not find the subject "%s" for the #[IsGranted] attribute. Try adding a "$%s" argument to your controller method.', $subjectRef, $subjectRef));
                 } else {
-                    $subject = $arguments[$subjectRef];
+                    $subject = $this->getIsGrantedSubject($subjectRef, $arguments);
                 }
             }
 
@@ -81,15 +78,37 @@ class IsGrantedAttributeListener implements EventSubscriberInterface
         return [KernelEvents::CONTROLLER_ARGUMENTS => ['onKernelControllerArguments', 10]];
     }
 
+    private function getIsGrantedSubject(string|Expression $subjectRef, array $arguments): mixed
+    {
+        if ($subjectRef instanceof Expression) {
+            $this->expressionLanguage ??= new ExpressionLanguage();
+
+            return $this->expressionLanguage->evaluate($subjectRef, [
+                'args' => $arguments,
+            ]);
+        }
+
+        if (!\array_key_exists($subjectRef, $arguments)) {
+            throw new RuntimeException(sprintf('Could not find the subject "%s" for the #[IsGranted] attribute. Try adding a "$%s" argument to your controller method.', $subjectRef, $subjectRef));
+        }
+
+        return $arguments[$subjectRef];
+    }
+
     private function getIsGrantedString(IsGranted $isGranted): string
     {
-        $processValue = fn ($value) => sprintf('"%s"', $value);
+        $processValue = fn ($value) => sprintf($value instanceof Expression ? 'new Expression("%s")' : '"%s"', $value);
 
         $argsString = $processValue($isGranted->attribute);
 
         if (null !== $subject = $isGranted->subject) {
-            $subject = array_map($processValue, (array) $subject);
-            $argsString .= ', '.(1 === \count($subject) ? reset($subject) : '['.implode(', ', $subject).']');
+            $subject = !\is_array($subject) ? $processValue($subject) : array_map(function ($key, $value) use ($processValue) {
+                $value = $processValue($value);
+
+                return \is_string($key) ? sprintf('"%s" => %s', $key, $value) : $value;
+            }, array_keys($subject), $subject);
+
+            $argsString .= ', '.(!\is_array($subject) ? $subject : '['.implode(', ', $subject).']');
         }
 
         return $argsString;